chore(fields-type): rename import
